moved hasScope method to token-utils
