+++ lib/token-manager/token-manager.js
@@ -21,135 +21,123 @@ const ServiceUtil = require('../utils/service-util');
 
 
 function TokenManager(options) {
-  logger.debug("Initializing token manager");
-  this.serviceConfig = new ServiceUtil.loadConfig('TokenManager', [
-	constants.CLIENT_ID,
-	constants.SECRET,
-	constants.OAUTH_SERVER_URL
-  ], options);
+	logger.debug("Initializing token manager");
+	this.serviceConfig = new ServiceUtil.loadConfig('TokenManager', [
+		constants.CLIENT_ID,
+		constants.SECRET,
+		constants.OAUTH_SERVER_URL
+	], options);
 }
 
 TokenManager.prototype.getCustomIdentityTokens = function (jwsToken, scopes = []) {
-  const clientId = this.serviceConfig.getClientId();
-  const secret = this.serviceConfig.getSecret();
-  const scope = scopes.join(' ');
-  const tokenEndpoint = this.serviceConfig.getOAuthServerUrl() + constants.TOKEN_PATH;
-  
-  
-  return getTokens(tokenEndpoint, clientId, secret, constants.CUSTOM_IDENTITY_GRANT_TYPE, scope, jwsToken)
-	.then((tokenResponse) => {
-	  if (!tokenResponse) {
-		logger.error('Unable to parse token response');
-		throw Error('Unable to parse token response');
-	  }
-	  
-	  const accessToken = tokenResponse['access_token'];
-	  const identityToken = tokenResponse['id_token'];
-	  const tokenType = tokenResponse['token_type'];
-	  const expiresIn = tokenResponse['expires_in'];
-	  
-	  return Promise.all([
-		validateToken('access', accessToken, this.serviceConfig),
-		validateToken('identity', identityToken, this.serviceConfig)
-	  ]).then(() => ({
-		accessToken,
-		identityToken,
-		tokenType,
-		expiresIn
-	  }));
-	});
+	const clientId = this.serviceConfig.getClientId();
+	const secret = this.serviceConfig.getSecret();
+	const scope = scopes.join(' ');
+	const tokenEndpoint = this.serviceConfig.getOAuthServerUrl() + constants.TOKEN_PATH;
+
+
+	return getTokens(tokenEndpoint, clientId, secret, constants.CUSTOM_IDENTITY_GRANT_TYPE, scope, jwsToken)
+		.then((tokenResponse) => {
+			if (!tokenResponse) {
+				logger.error('Unable to parse token response');
+				throw Error('Unable to parse token response');
+			}
+
+			const accessToken = tokenResponse['access_token'];
+			const identityToken = tokenResponse['id_token'];
+			const tokenType = tokenResponse['token_type'];
+			const expiresIn = tokenResponse['expires_in'];
+
+			return Promise.all([
+				validateToken('access', accessToken, this.serviceConfig),
+				validateToken('identity', identityToken, this.serviceConfig)
+			]).then(() => ({
+				accessToken,
+				identityToken,
+				tokenType,
+				expiresIn
+			}));
+		});
 };
 
 TokenManager.prototype.getApplicationIdentityToken = function () {
-  
-  const clientId = this.serviceConfig.getClientId();
-  const secret = this.serviceConfig.getSecret();
-  const tokenEndPoint = this.serviceConfig.getOAuthServerUrl() + constants.TOKEN_PATH;
-  
-  return getTokens(tokenEndPoint, clientId, secret, constants.APP_TO_APP_GRANT_TYPE)
-	.then((tokenResponse) => {
-	  if (!tokenResponse) {
-		logger.error('Unable to parse token response');
-		throw Error('Unable to parse token response');
-	  }
-	  
-	  const accessToken = tokenResponse['access_token'];
-	  const tokenType = tokenResponse['token_type'];
-	  const expiresIn = tokenResponse['expires_in'];
-	  
-	  return validateToken('access', accessToken, this.serviceConfig)
-		.then(() => ({
-		  accessToken,
-		  tokenType,
-		  expiresIn
-		}));
-	});
+
+	const clientId = this.serviceConfig.getClientId();
+	const secret = this.serviceConfig.getSecret();
+	const tokenEndPoint = this.serviceConfig.getOAuthServerUrl() + constants.TOKEN_PATH;
+
+	return getTokens(tokenEndPoint, clientId, secret, constants.APP_TO_APP_GRANT_TYPE)
+		.then((tokenResponse) => {
+			if (!tokenResponse) {
+				logger.error('Unable to parse token response');
+				throw Error('Unable to parse token response');
+			}
+
+			const accessToken = tokenResponse['access_token'];
+			const tokenType = tokenResponse['token_type'];
+			const expiresIn = tokenResponse['expires_in'];
+
+			return validateToken('access', accessToken, this.serviceConfig)
+				.then(() => ({
+					accessToken,
+					tokenType,
+					expiresIn
+				}));
+		});
 };
 
 function getTokens(tokenEndpoint, clientId, secret, grant_type, scope, assertion) {
-  return new Promise((resolve, reject) => {
-	request({
-	  method: 'POST',
-	  url: tokenEndpoint,
-	  auth: {
-		username: clientId,
-		password: secret
-	  },
-	  form: {
-		grant_type,
-		scope,
-		assertion
-	  }
-	}, (error, response, body) => {
-	  if (error) {
-		logger.error('Failed to obtain tokens. Error: ', error);
-		reject(error);
-	  } else if (response.statusCode === 200) {
-		resolve(JSON.parse(body));
-	  } else {
-		if (response.statusCode === 401 || response.statusCode === 403) {
-		  logger.error('Unauthorized. Error: ', JSON.parse(body).error_description);
-		  reject(new Error('Unauthorized'));
-		} else if (response.statusCode === 404) {
-		  logger.error('Not found');
-		  reject(new Error("Not found"));
-		} else if (response.statusCode === 400) {
-		  logger.error('Failed to obtain tokens. Error: ', JSON.parse(body).error_description);
-		  reject(new Error('Failed to obtain tokens'));
-		} else {
-		  logger.error(JSON.parse(body));
-		  reject(new Error("Unexpected error"));
-		}
-	  }
+	return new Promise((resolve, reject) => {
+		request({
+			method: 'POST',
+			url: tokenEndpoint,
+			auth: {
+				username: clientId,
+				password: secret
+			},
+			form: {
+				grant_type,
+				scope,
+				assertion
+			}
+		}, (error, response, body) => {
+			if (error) {
+				logger.error('Failed to obtain tokens. Error: ', error);
+				reject(error);
+			} else if (response.statusCode === 200) {
+				resolve(JSON.parse(body));
+			} else {
+				if (response.statusCode === 401 || response.statusCode === 403) {
+					logger.error('Unauthorized. Error: ', JSON.parse(body).error_description);
+					reject(new Error('Unauthorized'));
+				} else if (response.statusCode === 404) {
+					logger.error('Not found');
+					reject(new Error("Not found"));
+				} else if (response.statusCode === 400) {
+					logger.error('Failed to obtain tokens. Error: ', JSON.parse(body).error_description);
+					reject(new Error('Failed to obtain tokens'));
+				} else {
+					logger.error(JSON.parse(body));
+					reject(new Error("Unexpected error"));
+				}
+			}
+		});
 	});
-  });
 }
 
 function validateToken(tokenType, token, serviceConfig) {
-  logger.debug(`Validating ${tokenType} token`);
-  return TokenUtil.decodeAndValidate(token, serviceConfig.getOAuthServerUrl()).then((validatedToken) => {
-	if (!validatedToken) {
-	  logger.error(`Invalid ${tokenType} token`);
-	  throw Error(`Invalid ${tokenType} token`);
-	}
-	
-	return TokenUtil.validateIssAndAud(validatedToken, serviceConfig).catch(error => {
-	  logger.error(error);
-	  throw Error(`${tokenType} token has invalid claims`);
-	});
-  });
-}
+	logger.debug(`Validating ${tokenType} token`);
+	return TokenUtil.decodeAndValidate(token, serviceConfig.getOAuthServerUrl()).then((validatedToken) => {
+		if (!validatedToken) {
+			logger.error(`Invalid ${tokenType} token`);
+			throw Error(`Invalid ${tokenType} token`);
+		}
 
-function hasScope(req, requiredScopes) {
-  const requiredScopesArray = requiredScopes.split(" ");
-  const existingScopesArray = req.session[WebAppStrategy.AUTH_CONTEXT].accessTokenPayload.scope.split(" ");
-  
-  for (const requiredScope in requiredScopesArray) {
-	if (!existingScopesArray.includes(requiredScope)) {
-	  return false;
-	}
-  }
-  return true;
+		return TokenUtil.validateIssAndAud(validatedToken, serviceConfig).catch(error => {
+			logger.error(error);
+			throw Error(`${tokenType} token has invalid claims`);
+		});
+	});
 }
 
 module.exports = TokenManager;
